+++ knut/demo/milloneup/sys-milloneup.cpp
@@ -155,10 +155,10 @@ static inline double d_scf_x0x1( double phi, double x0, double x1 )
 extern "C"
 {
 
-int sys_ndim(){ return 2; }
-int sys_npar(){ return 3; }
-int sys_ntau(){ return 2; }
-int sys_nderi(){ return 2; }
+size_t sys_ndim(){ return 2; }
+size_t sys_npar(){ return 3; }
+size_t sys_ntau(){ return 2; }
+size_t sys_nderi(){ return 2; }
 
 void sys_tau( KNVector& out, double t, const KNVector& par )
 {
@@ -166,7 +166,7 @@ void sys_tau( KNVector& out, double t, const KNVector& par )
   out(1) = par(0);
 }
 
-void sys_dtau( KNVector& out, double t, const KNVector& par, int vp )
+void sys_dtau( KNVector& out, double t, const KNVector& par, size_t vp )
 {
 	switch( vp )
 	{
@@ -198,7 +198,7 @@ void sys_rhs( KNVector& out, double t, const KNMatrix& x, const KNVector& par )
 }
 
 void sys_deri( KNMatrix &out, double t, const KNMatrix& x, const KNVector& par, 
-	       int nx, const int* vx, int np, const int* vp, const KNMatrix& vv )
+	       size_t nx, const size_t* vx, size_t np, const size_t* vp, const KNMatrix& vv )
 {
   if( (t < 0)||(t > 1) ){ std::cout << "deri: t is not element of the interval\n"; }
   double g;
