+++ knut/demo/vanderpol/sys-vanderpol.cpp
@@ -10,10 +10,10 @@
 extern "C"
 {
 
-int sys_ndim(){ return 2; }
-int sys_npar(){ return 2; }
-int sys_ntau(){ return 1; }
-int sys_nderi(){ return 0; }
+size_t sys_ndim(){ return 2; }
+size_t sys_npar(){ return 2; }
+size_t sys_ntau(){ return 1; }
+size_t sys_nderi(){ return 0; }
 
 
 void sys_tau( KNVector& out, double t, const KNVector& par )
@@ -21,7 +21,7 @@ void sys_tau( KNVector& out, double t, const KNVector& par )
   out(0) = 0.0;
 }
 
-void sys_dtau( KNVector& out, double t, const KNVector& par, int vp )
+void sys_dtau( KNVector& out, double t, const KNVector& par, size_t vp )
 {
 	out(0) = 0.0;
 }
@@ -33,7 +33,7 @@ void sys_rhs( KNVector& out, double t, const KNMatrix& x, const KNVector& par )
 }
 
 void sys_deri( KNMatrix &out, double t, const KNMatrix& x, const KNVector& par, 
-	       int nx, const int* vx, int np, const int* vp, const KNMatrix& vv )
+	       size_t nx, const size_t* vx, size_t np, const size_t* vp, const KNMatrix& vv )
 {
 
 }
